Add listener to files which was created for the tasks
